services/vpn-admin-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.14
PyYAML==6.0.2
